% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_genres.R
\name{harmonize_genres}
\alias{harmonize_genres}
\title{Harmonize genres (ENB)}
\usage{
harmonize_genres(genres, lang = "ENG")
}
\arguments{
\item{genres}{the data.table format genre_by_row table}
}
\description{
This function harmonizes genres in ENB
}
\examples{
harmonize_genres()
see https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/ on installation
}
\keyword{genres}
