% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_places.R
\name{harmonize_places}
\alias{harmonize_places}
\title{Harmonize places (ENB)}
\usage{
harmonize_places(works, fuzzy = T)
}
\arguments{
\item{genres}{the data.table format work_by_row table}
}
\description{
This function harmonizes places in ENB
}
\examples{
harmonize_places()
see https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/ on installation


}
\keyword{places}
