% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_publishers.R
\name{harmonize_publishers}
\alias{harmonize_publishers}
\title{Harmonize publishers (ENB)}
\usage{
harmonize_publishers(works)
}
\arguments{
\item{genres}{the data.table format work_by_row table}
}
\description{
This function harmonizes publishers in ENB
}
\examples{
harmonize_publishers()
see https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/ on installation


}
\keyword{publishers}
