% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marc2tidy_function.R
\name{marc2tidy}
\alias{marc2tidy}
\title{marc 2 tidy function (ENB)}
\usage{
marc2tidy(file3)
}
\arguments{
\item{genres}{the data.table format work_by_row table}
}
\description{
This function converts a long tsv in marc format into a tidy wide table with extracted information.
}
\examples{
harmonize_publishers()
see https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/ on installation


}
\keyword{publishers}
